moved condition and expression subclasses to internal package
