+++ src/main/java/net/itarray/automotion/internal/properties/BinaryExpression.java
@@ -1,7 +1,8 @@
-package net.itarray.automotion.validation.properties;
+package net.itarray.automotion.internal.properties;
 
 import net.itarray.automotion.internal.geometry.Direction;
 import net.itarray.automotion.internal.properties.Context;
+import net.itarray.automotion.validation.properties.Expression;
 
 import java.util.function.BiFunction;
