+++ src/main/java/net/itarray/automotion/internal/properties/ElementPropertyExpression.java
@@ -1,10 +1,10 @@
-package net.itarray.automotion.validation.properties;
+package net.itarray.automotion.internal.properties;
 
 import net.itarray.automotion.internal.UIElement;
 import net.itarray.automotion.internal.geometry.Direction;
 import net.itarray.automotion.internal.geometry.ExtendGiving;
 import net.itarray.automotion.internal.geometry.Group;
-import net.itarray.automotion.internal.properties.Context;
+import net.itarray.automotion.validation.properties.Expression;
 
 import java.util.function.Function;
