+++ src/main/java/net/itarray/automotion/internal/properties/PercentReference.java
@@ -1,0 +1,5 @@
+package net.itarray.automotion.internal.properties;
+
+public enum  PercentReference {
+    PAGE;
+}
