+++ src/main/java/net/itarray/automotion/validation/ChunkUIElementValidator.java
@@ -53,12 +53,12 @@ public interface ChunkUIElementValidator {
     ChunkUIElementValidator dontDrawMap();
 
     /**
-     * @deprecated As of release 2.0, replaced by {@link net.itarray.automotion.validation.properties.Expression#percent(int, net.itarray.automotion.validation.properties.PercentReference)}
+     * @deprecated As of release 2.0, replaced by {@link net.itarray.automotion.validation.properties.Expression#percent(int, net.itarray.automotion.internal.properties.PercentReference)}
      */
     @Deprecated
     ChunkUIElementValidator changeMetricsUnitsTo(util.validator.ResponsiveUIValidator.Units units);
     /**
-     * @deprecated As of release 2.0, replaced by {@link net.itarray.automotion.validation.properties.Expression#percent(int, net.itarray.automotion.validation.properties.PercentReference)}
+     * @deprecated As of release 2.0, replaced by {@link net.itarray.automotion.validation.properties.Expression#percent(int, net.itarray.automotion.internal.properties.PercentReference)}
      */
     @Deprecated
     ChunkUIElementValidator changeMetricsUnitsTo(Units units);
