+++ src/main/java/net/itarray/automotion/validation/ResponsiveUIValidator.java
@@ -115,7 +115,7 @@ public class ResponsiveUIValidator {
     }
 
     /**
-     * @deprecated As of release 2.0, replaced by {@link net.itarray.automotion.validation.properties.Expression#percent(int, net.itarray.automotion.validation.properties.PercentReference)}
+     * @deprecated As of release 2.0, replaced by {@link net.itarray.automotion.validation.properties.Expression#percent(int, net.itarray.automotion.internal.properties.PercentReference)}
      */
     @Deprecated
     public ResponsiveUIValidator changeMetricsUnitsTo(net.itarray.automotion.validation.Units units) {
@@ -124,7 +124,7 @@ public class ResponsiveUIValidator {
     }
 
     /**
-     * @deprecated As of release 2.0, replaced by {@link net.itarray.automotion.validation.properties.Expression#percent(int, net.itarray.automotion.validation.properties.PercentReference)}
+     * @deprecated As of release 2.0, replaced by {@link net.itarray.automotion.validation.properties.Expression#percent(int, net.itarray.automotion.internal.properties.PercentReference)}
      */
     @Deprecated
     public ResponsiveUIValidator changeMetricsUnitsTo(util.validator.ResponsiveUIValidator.Units units) {
@@ -180,7 +180,7 @@ public class ResponsiveUIValidator {
     }
 
     /**
-     * @deprecated As of release 2.0, replaced by {@link net.itarray.automotion.validation.properties.Expression#percent(int, net.itarray.automotion.validation.properties.PercentReference)}
+     * @deprecated As of release 2.0, replaced by {@link net.itarray.automotion.validation.properties.Expression#percent(int, net.itarray.automotion.internal.properties.PercentReference)}
      */
     public static class Units {
         public static net.itarray.automotion.validation.Units PX = net.itarray.automotion.validation.Units.PX;
