+++ src/main/java/net/itarray/automotion/validation/Units.java
@@ -1,7 +1,7 @@
 package net.itarray.automotion.validation;
 
 /**
- * @deprecated As of release 2.0, replaced by {@link net.itarray.automotion.validation.properties.Expression#percent(int, net.itarray.automotion.validation.properties.PercentReference)}
+ * @deprecated As of release 2.0, replaced by {@link net.itarray.automotion.validation.properties.Expression#percent(int, net.itarray.automotion.internal.properties.PercentReference)}
  */
 @Deprecated
 public enum Units {
