--- src/main/java/net/itarray/automotion/validation/properties/PercentReference.java
@@ -1,5 +1,0 @@
-package net.itarray.automotion.validation.properties;
-
-public enum  PercentReference {
-    PAGE;
-}
