+++ src/main/java/util/validator/ResponsiveUIValidator.java
@@ -176,7 +176,7 @@ public class ResponsiveUIValidator {
     }
 
     /**
-     * @deprecated As of release 2.0, replaced by {@link net.itarray.automotion.validation.properties.Expression#percent(int, net.itarray.automotion.validation.properties.PercentReference)}
+     * @deprecated As of release 2.0, replaced by {@link net.itarray.automotion.validation.properties.Expression#percent(int, net.itarray.automotion.internal.properties.PercentReference)}
      */
     @Deprecated
     public enum Units {
