+++ src/test/java/net/itarray/automotion/tests/properties/ElementPropertyExpressionTest.java
@@ -4,7 +4,7 @@ import net.itarray.automotion.internal.UIElement;
 import net.itarray.automotion.internal.geometry.Direction;
 import net.itarray.automotion.internal.geometry.Rectangle;
 import net.itarray.automotion.internal.geometry.Scalar;
-import net.itarray.automotion.validation.properties.ElementPropertyExpression;
+import net.itarray.automotion.internal.properties.ElementPropertyExpression;
 import net.itarray.automotion.validation.properties.Expression;
 import org.junit.Before;
 import org.junit.Test;
